requirements.txt
Django==2.0.2
pytz==2018.3
certifi==2018.1.18
chardet==3.0.4
Django==2.0.2
django-filter==1.1.0
djangorestframework==3.7.7
httpie==0.9.9
idna==2.6
Markdown==2.6.11
Pygments==2.2.0
pytz==2018.3
requests==2.18.4
urllib3==1.22
